# green-eco-bot 2/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: admin
    description: Admin panel operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Dashboard statistics
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: job
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: number
        - name: offset
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"

  /admin/users/{telegramId}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get user details
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update user balance or lawfulness
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/catalog:
    get:
      operationId: listCatalogItems
      tags: [admin]
      summary: List all marketplace catalog items
      responses:
        "200":
          description: Catalog items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CatalogItem"
    post:
      operationId: createCatalogItem
      tags: [admin]
      summary: Create a new catalog item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatalogItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogItem"

  /admin/catalog/{id}:
    patch:
      operationId: updateCatalogItem
      tags: [admin]
      summary: Update a catalog item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatalogItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCatalogItem
      tags: [admin]
      summary: Delete a catalog item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/activity:
    get:
      operationId: getAdminActivity
      tags: [admin]
      summary: Recent bot activity feed
      parameters:
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Activity items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  /admin/market:
    get:
      operationId: listMarketListings
      tags: [admin]
      summary: Active drug market listings
      responses:
        "200":
          description: Market listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketListingItem"

  /admin/market/{id}/deactivate:
    post:
      operationId: deactivateMarketListing
      tags: [admin]
      summary: Deactivate a market listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    DeleteResult:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok

    AdminStats:
      type: object
      properties:
        totalUsers:
          type: number
        activeWorkers:
          type: number
        totalBalance:
          type: number
        totalLawfulness:
          type: number
        jobBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/JobCount"
        topEarners:
          type: array
          items:
            $ref: "#/components/schemas/TopEarner"
        unlockedPolice:
          type: number
        unlockedDealers:
          type: number
        pendingShifts:
          type: number
        pendingMissions:
          type: number
      required:
        - totalUsers
        - activeWorkers
        - totalBalance
        - totalLawfulness
        - jobBreakdown
        - topEarners
        - unlockedPolice
        - unlockedDealers
        - pendingShifts
        - pendingMissions

    JobCount:
      type: object
      properties:
        job:
          type: string
        count:
          type: number
      required:
        - job
        - count

    TopEarner:
      type: object
      properties:
        telegramId:
          type: number
        username:
          type: ["string", "null"]
        firstName:
          type: string
        balance:
          type: number
      required:
        - telegramId
        - username
        - firstName
        - balance

    AdminUser:
      type: object
      properties:
        telegramId:
          type: number
        username:
          type: ["string", "null"]
        firstName:
          type: string
        balance:
          type: number
        lawfulness:
          type: number
        currentJob:
          type: ["string", "null"]
        policeUnlocked:
          type: boolean
        dealerUnlocked:
          type: boolean
        createdAt:
          type: string
      required:
        - telegramId
        - username
        - firstName
        - balance
        - lawfulness
        - currentJob
        - policeUnlocked
        - dealerUnlocked
        - createdAt

    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: number
      required:
        - users
        - total

    UserDetail:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AdminUser"
        policeData:
          $ref: "#/components/schemas/PoliceInfo"
        dealerData:
          $ref: "#/components/schemas/DealerInfo"
        inventoryValue:
          type: number
        purchasedItems:
          type: number
      required:
        - user
        - policeData
        - dealerData
        - inventoryValue
        - purchasedItems

    PoliceInfo:
      type: object
      properties:
        rank:
          type: ["string", "null"]
        missionsCompleted:
          type: number
        totalMissions:
          type: number
        stationLevel:
          type: number
      required:
        - rank
        - missionsCompleted
        - totalMissions
        - stationLevel

    DealerInfo:
      type: object
      properties:
        farmLevel:
          type: number
        workersCount:
          type: number
        lastHarvest:
          type: ["string", "null"]
      required:
        - farmLevel
        - workersCount
        - lastHarvest

    UserUpdate:
      type: object
      properties:
        balance:
          type: number
        lawfulness:
          type: number

    CatalogItem:
      type: object
      properties:
        id:
          type: number
        botNetwork:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        emoji:
          type: string
        isAvailable:
          type: boolean
      required:
        - id
        - botNetwork
        - name
        - description
        - price
        - category
        - emoji
        - isAvailable

    CatalogItemInput:
      type: object
      properties:
        botNetwork:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        emoji:
          type: string
      required:
        - botNetwork
        - name
        - description
        - price
        - category
        - emoji

    CatalogItemUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        emoji:
          type: string
        isAvailable:
          type: boolean

    ActivityItem:
      type: object
      properties:
        type:
          type: string
        description:
          type: string
        userId:
          type: number
        username:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        happenedAt:
          type: string
      required:
        - type
        - description
        - userId
        - username
        - amount
        - happenedAt

    MarketListingItem:
      type: object
      properties:
        id:
          type: number
        sellerTelegramId:
          type: number
        sellerUsername:
          type: ["string", "null"]
        substanceKey:
          type: string
        substanceName:
          type: string
        substanceClass:
          type: string
        quantity:
          type: number
        pricePerUnit:
          type: number
        totalValue:
          type: number
        listedAt:
          type: string
      required:
        - id
        - sellerTelegramId
        - sellerUsername
        - substanceKey
        - substanceName
        - substanceClass
        - quantity
        - pricePerUnit
        - totalValue
        - listedAt